chore: converted files to typescript
